+++ src/main/java/net/volcano/jdacommands/model/command/arguments/implementation/codec/LongCodec.java
@@ -17,8 +17,8 @@ public class LongCodec extends Codec<Long> {
 		var arg = data.parameter.getAnnotation(Arg.class);
 		
 		if (arg != null) {
-			var aMin = arg.min() == Double.MIN_VALUE ? null : (long) arg.min();
-			var aMax = arg.max() == Double.MAX_VALUE ? null : (long) arg.max();
+			var aMin = arg.min() == Long.MIN_VALUE ? null : (long) arg.min();
+			var aMax = arg.max() == Long.MAX_VALUE ? null : (long) arg.max();
 			
 			builder.min(aMin);
 			builder.max(aMax);
